+++ ChangeLog.md
@@ -14,6 +14,8 @@ Eventum Issue Tracking System
 - drop unused Misc::getInput (5b97d89)
 - drop prompt and getInputLine from Misc (83efe3c)
 - drop unused Misc::collect (51eab96)
+- drop table prefix support (#166, @glensc)
+- drop old `APP_SQL_` constants support, support upgrade from earlier than 3.0 versions dropped
 
 2016-04-28, Version [3.1.0]
 ----------------------------